+++ tests/TestCase/FilesystemFunctionsTest.php
@@ -15,7 +15,6 @@ declare(strict_types=1);
 
 namespace Tools\Test;
 
-use PHPUnit\Framework\Error\Deprecated;
 use Tools\TestSuite\TestCase;
 
 /**
@@ -179,22 +178,6 @@ class FilesystemFunctionsTest extends TestCase
         ] as $url) {
             $this->assertEquals('sql.gz', get_extension($url));
         }
-    }
-
-    /**
-     * Test for `is_absolute()` global function
-     * @test
-     */
-    public function testIsAbsolute()
-    {
-        $errorReporting = error_reporting(E_ALL & ~E_USER_DEPRECATED);
-        $this->assertTrue(is_absolute(DS . 'path' . DS));
-        $this->assertFalse(is_absolute('path' . DS));
-        error_reporting($errorReporting);
-
-        $this->expectException(Deprecated::class);
-        $this->expectExceptionMessage('`is_absolute()` function is deprecated. Use `Filesystem::isAbsolutePath()` instead');
-        is_absolute('path' . DS);
     }
 
     /**